renamed to sdk-react-native, updated readme
