--- packages/react-native-sessions/src/index.ts
@@ -1,3 +1,0 @@
-export type { SessionConfig, Session, User } from "./contexts/SessionContext";
-export { SessionContext, SessionProvider } from "./contexts/SessionContext";
-export { useSession } from "./hooks/use-session";
