+++ packages/sdk-react-native/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@turnkey/react-native-sessions",
+  "name": "@turnkey/sdk-react-native",
   "version": "0.1.0",
-  "description": "Session management for React Native using Turnkey",
+  "description": "React Native SDK",
   "main": "./dist/index.js",
   "module": "./dist/index.mjs",
   "types": "./dist/index.d.ts",
@@ -15,11 +15,7 @@
   },
   "license": "Apache-2.0",
   "keywords": [
-    "turnkey",
-    "session",
-    "react-native",
-    "expo",
-    "secure-storage"
+    "turnkey"
   ],
   "author": {
     "name": "Turnkey",
@@ -32,7 +28,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/tkhq/sdk.git",
-    "directory": "packages/react-native-sessions"
+    "directory": "packages/sdk-react-native"
   },
   "files": [
     "dist/",
