+++ packages/sdk-react-native/src/constant.ts
